+++ lib/src/ast/css/style_rule.dart
@@ -2,6 +2,8 @@
 // MIT-style license that can be found in the LICENSE file or at
 // https://opensource.org/licenses/MIT.
 
+import 'package:meta/meta.dart';
+
 import '../selector.dart';
 import 'node.dart';
 
@@ -16,4 +18,10 @@ abstract interface class CssStyleRule implements CssParentNode {
 
   /// The selector for this rule, before any extensions were applied.
   SelectorList get originalSelector;
+
+  /// Whether this style rule was originally defined in a plain CSS stylesheet.
+  ///
+  /// :nodoc:
+  @internal
+  bool get fromPlainCss;
 }