+++ lib/src/visitor/evaluate.dart
@@ -5,7 +5,7 @@
 // DO NOT EDIT. This file was generated from async_evaluate.dart.
 // See tool/grind/synchronize.dart for details.
 //
-// Checksum: 7351193aa9229e1434c09a2cbc9fa596cd924901
+// Checksum: 05cb957cd0c7698d8ad648f31d862dc91f0daa7b
 //
 // ignore_for_file: unused_import
 
@@ -2001,16 +2001,32 @@ final class _EvaluateVisitor
     }
 
     var parsedSelector = SelectorList.parse(selectorText,
-            interpolationMap: selectorMap,
-            allowParent: !_stylesheet.plainCss,
-            allowPlaceholder: !_stylesheet.plainCss,
-            logger: _logger)
-        .resolveParentSelectors(_styleRuleIgnoringAtRoot?.originalSelector,
-            implicitParent: !_atRootExcludingStyleRule);
+        interpolationMap: selectorMap,
+        plainCss: _stylesheet.plainCss,
+        logger: _logger);
+
+    var nest = !(_styleRule?.fromPlainCss ?? false);
+    if (nest) {
+      if (_stylesheet.plainCss) {
+        for (var complex in parsedSelector.components) {
+          if (complex.leadingCombinators case [var first, ...]
+              when _stylesheet.plainCss) {
+            throw _exception(
+                "Top-level leading combinators aren't allowed in plain CSS.",
+                first.span);
+          }
+        }
+      }
+
+      parsedSelector = parsedSelector.nestWithin(
+          _styleRuleIgnoringAtRoot?.originalSelector,
+          implicitParent: !_atRootExcludingStyleRule,
+          preserveParentSelectors: _stylesheet.plainCss);
+    }
 
     var selector = _extensionStore.addSelector(parsedSelector, _mediaQueries);
     var rule = ModifiableCssStyleRule(selector, node.span,
-        originalSelector: parsedSelector);
+        originalSelector: parsedSelector, fromPlainCss: _stylesheet.plainCss);
     var oldAtRootExcludingStyleRule = _atRootExcludingStyleRule;
     _atRootExcludingStyleRule = false;
     _withParent(rule, () {
@@ -2020,7 +2036,7 @@ final class _EvaluateVisitor
         }
       });
     },
-        through: (node) => node is CssStyleRule,
+        through: nest ? (node) => node is CssStyleRule : null,
         scopeWhen: node.hasDeclarations);
     _atRootExcludingStyleRule = oldAtRootExcludingStyleRule;
 
@@ -2038,13 +2054,15 @@ final class _EvaluateVisitor
               complex.span.trimRight(),
               Deprecation.bogusCombinators);
         } else if (complex.leadingCombinators.isNotEmpty) {
-          _warn(
-              'The selector "${complex.toString().trim()}" is invalid CSS.\n'
-              'This will be an error in Dart Sass 2.0.0.\n'
-              '\n'
-              'More info: https://sass-lang.com/d/bogus-combinators',
-              complex.span.trimRight(),
-              Deprecation.bogusCombinators);
+          if (!_stylesheet.plainCss) {
+            _warn(
+                'The selector "${complex.toString().trim()}" is invalid CSS.\n'
+                'This will be an error in Dart Sass 2.0.0.\n'
+                '\n'
+                'More info: https://sass-lang.com/d/bogus-combinators',
+                complex.span.trimRight(),
+                Deprecation.bogusCombinators);
+          }
         } else {
           _warn(
               'The selector "${complex.toString().trim()}" is only valid for '
@@ -3354,12 +3372,15 @@ final class _EvaluateVisitor
     }
 
     var styleRule = _styleRule;
-    var originalSelector = node.selector.resolveParentSelectors(
-        styleRule?.originalSelector,
-        implicitParent: !_atRootExcludingStyleRule);
+    var nest = !(_styleRule?.fromPlainCss ?? false);
+    var originalSelector = nest
+        ? node.selector.nestWithin(styleRule?.originalSelector,
+            implicitParent: !_atRootExcludingStyleRule,
+            preserveParentSelectors: node.fromPlainCss)
+        : node.selector;
     var selector = _extensionStore.addSelector(originalSelector, _mediaQueries);
     var rule = ModifiableCssStyleRule(selector, node.span,
-        originalSelector: originalSelector);
+        originalSelector: originalSelector, fromPlainCss: node.fromPlainCss);
     var oldAtRootExcludingStyleRule = _atRootExcludingStyleRule;
     _atRootExcludingStyleRule = false;
     _withParent(rule, () {
@@ -3368,7 +3389,7 @@ final class _EvaluateVisitor
           child.accept(this);
         }
       });
-    }, through: (node) => node is CssStyleRule, scopeWhen: false);
+    }, through: nest ? (node) => node is CssStyleRule : null, scopeWhen: false);
     _atRootExcludingStyleRule = oldAtRootExcludingStyleRule;
 
     if (_parent.children case [..., var lastChild] when styleRule == null) {